+++ src/tests/test_cases/unicode_literals.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-"""This is a module."""
+"""A module."""
 
 from __future__ import unicode_literals
 from .expected import Expectation